pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wepa"
version = "0.1.0"
description = "Paper code for Watermarks for Language Models via Probabilistic Automata"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets==4.4.2",
    "ipykernel==7.1.0",
    "ipywidgets==8.1.7",
    "jupyter==1.0.0",
    "matplotlib==3.10.8",
    "nbconvert==7.16.6",
    "numba==0.63.1",
    "numpy==2.3.5",
    "pandas==2.3.3",
    "pyyaml==6.0.3",
    "scikit-learn==1.8.0",
    "sentencepiece==0.2.1",
    "torch==2.9.1",
    "transformers==4.57.3",
    "wandb==0.23.1",
]

[tool.setuptools]
include-package-data = true
packages = ["experiments", "experiments.runs", "watermarker"]

[tool.setuptools.package-dir]
experiments = "experiments"
watermarker = "src/watermarker"
